Refactor ModelViewerView and ScannerView for improved functionality and code clarity
- Integrated ModelViewerMixin into ModelViewerView to streamline model handling.
- Simplified rotation logic and removed unnecessary state variables in ModelViewerView.
- Updated ScannerView to enhance button visibility based on scanning state.
